# docs/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ingest/candidate-assets/{id}/preview": {
      "get": {
        "operationId": "IngestAdminController_candidatePreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
...
    "/api/supply/sources": {
      "get": {
        "operationId": "SupplyController_sources",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "货源与供应商"
        ]
      }
    },
    "/api/supply/sources/{id}": {
      "get": {
        "operationId": "SupplyController_source",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "货源与供应商"
        ]
      }
    },
    "/api/supply/sources/{id}/revisions": {
      "get": {
        "operationId": "SupplyController_revisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "货源与供应商"
        ]
      }
    },
...
    "/api/supply/offers/{id}/withdraw": {
      "post": {
        "operationId": "SupplyController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "货源与供应商"
        ]
      }
    },
...
  },
  "info": {
    "title": "ToMeBoutique Workbench API",
    "description": "",
    "version": "1.0.0-rc.14",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tome_session"
      }
    },
...
  }
}